chore: move pub-sub to redis folder
